# src/main/resources/static/openapi.yaml
...
openapi: 3.0.1
info:
  title: RSAPI
  description: |
    Backend Service for https://www.railway-stations.org/. 
    Sourcecode available at: https://github.com/RailwayStations/RSAPI.
  license:
    name: MIT
    url: https://github.com/RailwayStations/RSAPI/blob/master/LICENSE.txt
  version: "20221126.2100"
servers:
  - url: https://api.railway-stations.org/
...
paths:
  /photoStationById/{country}/{id}:
    get:
      summary: Get a railway station of a country by its id with all its photos
      parameters:
        - name: country
          in: path
          description: country code
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: id of the station
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoStations'
        404:
          description: Station not found
          content: { }
  /photoStationsByCountry/{country}:
    get:
      summary: List stations by country
      description: |
        Returns only the primary photo of each station.
        Stations can optionally be filtered by 'hasPhoto' and/or 'isActive' flag.
      parameters:
        - name: country
          in: path
          description: country code
          required: true
          schema:
            type: string
        - name: hasPhoto
          in: query
          description: filter by photo available/missing
          schema:
            type: boolean
        - name: isActive
          in: query
          description: filter on active/inactive stations
          schema:
            type: boolean
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoStations'
  /photoStationsByPhotographer/{photographer}:
    get:
      summary: List stations with photos by the given photographer
      description: |
        Returns only the photos of a station which belong to the photographer.
        Can return several photos per station, not necessarily the primary photo.
      parameters:
        - name: photographer
          in: path
          description: photographer name
          required: true
          schema:
            type: string
        - name: country
          in: query
          description: country code
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoStations'
  /photoStationsByRecentPhotoImports:
    get:
      summary: List stations with photos uploaded in the last 24h
      description: |
        Returns only the recently imported photos of a station, defined by the sinceHours parameter.
        Can return several photos per station, not necessarily the primary photo.
      parameters:
        - name: sinceHours
          in: query
          description: |
            defines the timeframe since when to look for recent photo uploads,
            default is 10h
          schema:
            type: integer
            format: int32
            default: 10
            minimum: 1
            maximum: 800
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoStations'
...
  /registration:
    post:
      summary: Register for an account
      requestBody:
        description: Registration information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterProfile'
        required: true
      responses:
        202:
          description: registration accepted
          content: { }
        400:
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
        409:
          description: conflict with existing name or email
          content:
            text/plain:
              schema:
                type: string
        422:
          description: unprocessable entity, Registration input document error
          content: { }
      x-codegen-request-body-name: registration
  /emailVerification/{token}:
    get:
      summary: Verify eMail-Address with verification token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: email successfully verified
          content:
            text/plain:
              schema:
                type: string
        404:
          description: token not found, verification failed
          content: { }
  /resendEmailVerification:
    post:
      summary: Sends a new eMail-Verification mail
      parameters:
        - $ref: '#/components/parameters/Authorization'
      responses:
        200:
          description: email successfully sent
          content: { }
  /resetPassword:
    post:
      summary: Requests a new password only
      parameters:
        - name: NameOrEmail
          in: header
          description: name or email of the user
          required: true
          schema:
            type: string
      responses:
        202:
          description: request accepted, email with new password will be sent
          content: { }
        400:
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
  /changePassword:
    post:
      summary: Change user password
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - name: New-Password
          in: header
          description: new password, deprecated use body message for proper encoding
          deprecated: true
          required: false
          schema:
            type: string
      requestBody:
        description: ChangePassword
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePassword'
        required: false
      responses:
        200:
          description: password changed
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                type: string
        400:
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
        404:
          description: User not found
          content: { }
  /myProfile:
    get:
      summary: Get the users profile
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/Upload-Token'
        - $ref: '#/components/parameters/Email'
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        401:
          description: authorization failed
          content: { }
        403:
          description: forbidden
          content: { }
    post:
      summary: Update the users profile.
      description: |
        In case the email has been changed, only the email is saved and a new password is send.
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/Upload-Token'
        - $ref: '#/components/parameters/Email'
      requestBody:
        description: Userprofile
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfile'
        required: true
      responses:
        200:
          description: ok
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                type: string
        400:
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
        401:
          description: authorization failed
          content: { }
        403:
          description: forbidden
          content: { }
        409:
          description: conflict with existing name or email
          content:
            text/plain:
              schema:
                type: string
      x-codegen-request-body-name: profile
...
components:
  parameters:
    Authorization:
      name: Authorization
      in: header
      description: |
        HTTP basic authorization
      required: false
      schema:
        type: string
    Upload-Token:
      name: Upload-Token
      in: header
      description: |
        DEPRECATED authentication via personal Upload-Token!
        Used together with Email header.
      deprecated: true
      required: false
      schema:
        type: string
    Email:
      name: Email
      in: header
      description: |
        DEPRECATED, use basic authorization! 
        Email or name of the user, used together with the Upload-Token.
      deprecated: true
      required: false
      schema:
        type: string
  schemas:
    PhotoStations:
      description: Stations with photos
      type: object
      properties:
        photoBaseUrl:
          type: string
          description: Base URL of all photos
          example: "https://api.railway-stations.org/photos/"
        licenses:
          type: array
          description: List of used licenses, might be empty if no photos available
          items:
            $ref: '#/components/schemas/PhotoLicense'
        photographers:
          type: array
          description: List of all photographers, might be empty if no photos available
          items:
            $ref: '#/components/schemas/Photographer'
        stations:
          type: array
          description: List of the stations
          items:
            $ref: '#/components/schemas/PhotoStation'
      required:
        - "photoBaseUrl"
        - "licenses"
        - "photographers"
        - "stations"
    PhotoLicense:
      description: License used by a photo
      type: object
      properties:
        id:
          type: string
          description: Unique id of the license
          example: "CC0"
        name:
          type: string
          description: Name of the license to display at the photo
          example: "CC0 1.0 Universell (CC0 1.0)"
        url:
          type: string
          format: uri
          description: URL of the license to link to from the photo
          example: "https://creativecommons.org/publicdomain/zero/1.0/"
      required:
        - "id"
        - "name"
        - "url"
    Photographer:
      description: The creator of a photo
      type: object
      properties:
        name:
          type: string
          description: Username of the photographer
        url:
          type: string
          format: uri
          description: Link to the photographers social media account or homepage
      required:
        - "name"
    PhotoStation:
      description: A station with its photos
      type: object
      properties:
        country:
          type: string
          description: 2 letter code of the country
          example: "uk"
        id:
          type: string
          description: Id of the station within the country
          example: "7054260"
        title:
          type: string
          description: Title of the station
          example: "London Victoria"
        lat:
          type: number
          format: double
          description: Latitude of the station
        lon:
          type: number
          format: double
          description: Longitute of the station
        shortCode:
          type: string
          description: Provider specific short code of the station, e.g. RIL100 or DS100 for german stations
        inactive:
          type: boolean
          description: Indicates if this station is inactive
          default: false
        photos:
          type: array
          description: Photos of the station. If more than one photo is given, the first one is the primary photo. List might be empty or only the primary photo provided.
          items:
            $ref: '#/components/schemas/Photo'
      required:
        - "country"
        - "id"
        - "title"
        - "lat"
        - "lon"
        - "photos"
    Photo:
      description: A photo of a station
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique id of a photo
        photographer:
          type: string
          description: Name of the photographer
        path:
          type: string
          description: URL path to the photo, to be used together with the photoBaseUrl
        createdAt:
          type: integer
          format: int64
          description: |
            Timestamp when the photo was created in the railway-stations
            database (Epoche milliseconds since 1.1.1970)
        license:
          type: string
          description: Id of the license used for this photo
        outdated:
          type: boolean
          description: Indicates if this photo is outdated
          default: false
      required:
        - "id"
        - "photographer"
        - "path"
        - "createdAt"
        - "license"
...
    License:
      description: the only accepted type is "CC0 1.0 Universell (CC0 1.0)", the others are listed for backward compatibility
      type: string
      enum:
        - CC0
        - CC0 1.0 Universell (CC0 1.0)
        - CC4
        - CC BY-SA 4.0
    UpdateProfile:
      description: User profile information
      type: object
      properties:
        nickname:
          type: string
        email:
          type: string
          format: email
        license:
          $ref: '#/components/schemas/License'
        photoOwner:
          type: boolean
        link:
          type: string
          format: uri
        anonymous:
          type: boolean
        sendNotifications:
          type: boolean
      required:
        - nickname
        - email
        - license
        - photoOwner
    RegisterProfile:
      description: User profile information
      type: object
      properties:
        nickname:
          type: string
        email:
          type: string
          format: email
        license:
          $ref: '#/components/schemas/License'
        photoOwner:
          type: boolean
        link:
          type: string
          format: uri
        anonymous:
          type: boolean
        sendNotifications:
          type: boolean
        newPassword:
          type: string
      required:
        - nickname
        - email
        - license
        - photoOwner
    Profile:
      description: User profile information
      type: object
      properties:
        nickname:
          type: string
        email:
          type: string
          format: email
        license:
          $ref: '#/components/schemas/License'
        photoOwner:
          type: boolean
        link:
          type: string
          format: uri
        anonymous:
          type: boolean
        admin:
          type: boolean
        emailVerified:
          type: boolean
        sendNotifications:
          type: boolean
      required:
        - nickname
        - license
        - photoOwner
...
    ChangePassword:
      type: object
      properties:
        newPassword:
          type: string
      required:
        - newPassword
...